split packaging and conda-build 3 updates

+++ cross-compilers/gcc/install-gfortran.sh
@@ -1,7 +1,11 @@
 set -e -x
 
-CHOST="x86_64-sarc-linux-gnu"
-_libdir="lib/gcc/$CHOST/$PKG_VERSION"
+
+CHOST="x86_64-${vendor}-linux-gnu"
+PACKAGE="gfortran"
+
+
+_libdir="libexec/gcc/$CHOST/$PKG_VERSION"
 pushd $SRC_DIR/.build/$CHOST/build/build-cc-gcc-final/
 
 # adapted from Arch install script from https://github.com/archlinuxarm/PKGBUILDs/blob/master/core/gcc/PKGBUILD
@@ -11,7 +15,9 @@ make -C $CHOST/libgomp prefix=$PREFIX install-nodist_fincludeHEADERS
 make -C gcc prefix=$PREFIX fortran.install-{common,man,info}
 install -Dm755 gcc/f951 $PREFIX/${_libdir}/f951
 
-ln -s $PREFIX/bin/$CHOST-gfortran $PREFIX/bin/f95
+if [ ! -e ${PREFIX}/bin/f95 ]; then
+    ln -s $PREFIX/bin/$CHOST-gfortran $PREFIX/bin/f95
+fi
 
 popd
 
@@ -20,7 +26,7 @@ install -Dm644 $SRC_DIR/.build/src/gcc-${PKG_VERSION}/COPYING.RUNTIME \
         ${PREFIX}/share/licenses/gcc-fortran/RUNTIME.LIBRARY.EXCEPTION
 
 mkdir -p $PREFIX/etc/conda/activate.d
-cp $RECIPE_DIR/activate-gfortran.sh $PREFIX/etc/conda/activate.d/compiler_linux-cos5-64-activate-gfortran.sh
+echo "export CHOST=${CHOST}" | cat - $RECIPE_DIR/activate-${PACKAGE}.sh > /tmp/out && mv /tmp/out $PREFIX/etc/conda/activate.d/activate-${PACKAGE}-${CHOST}.sh
 
 mkdir -p $PREFIX/etc/conda/deactivate.d
-cp $RECIPE_DIR/deactivate-gfortran.sh $PREFIX/etc/conda/deactivate.d/compiler_linux-cos5-64-deactivate-gfortran.sh
+echo "export CHOST=${CHOST}" | cat - $RECIPE_DIR/deactivate-${PACKAGE}.sh > /tmp/out && mv /tmp/out $PREFIX/etc/conda/deactivate.d/deactivate-${PACKAGE}-${CHOST}.sh
